add support for osx64

+++ waUtility.h
@@ -68,7 +68,7 @@ enum extract_option {
 #define EXTRACT_ALL	(EXTRACT_ALPHA|EXTRACT_DIGIT|EXTRACT_PUNCT|EXTRACT_SPACE|EXTRACT_HTML)
 
 /// 返回字符串HASH值，基于DJB HASH算法
-unsigned int string_hash( const string &str );
+size_t string_hash( const string &str );
 
 /// 全文词表替换，兼容GBK汉字
 string replace_text( const string &text, const map<string,string> &replace );
